log/pyproject.toml
[tool.poetry]
name = "log"
version = "0.1.0"
description = "High level logging application for MOL."
authors = ["Per Buer <per.buer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.49.0"
yoyo-migrations = "^7.0.1"
psycopg2 = "^2.8.4"
uvicorn = "^0.11.3"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-mock = "^2.0.0"
pytest-dependency = "^0.5.1"
requests = "^2.23.0"
python-dotenv = "^0.11.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
